+++ package.json
@@ -25,6 +25,8 @@
   },
   "dependencies": {
     "@dhis2/app-runtime": "^2.6.1",
-    "@dhis2/ui": "^5.7.2"
+    "@dhis2/ui": "^5.7.2",
+    "react-dnd": "^11.1.3",
+    "react-dnd-html5-backend": "^11.1.3"
   }
 }
